requirements.txt
Flask==3.0.2
psycopg==3.1.18
psycopg-binary==3.1.18
PyJWT==2.8.0
bcrypt==4.1.2
SQLAlchemy==2.0.28
pandas==2.1.4